pyproject.toml
[tool.poetry]
name = "facilito"
version = "0.5.0"
description = "simple api to scrape data from codigo facilito"
authors = ["ivansaul"]
readme = "README.md"
packages = [{include = "facilito", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest-playwright = "^0.5.2"
pydantic = "^2.10.0"
typer = "^0.13.0"
colorlog = "^6.9.0"
unidecode = "^1.3.8"
playwright-stealth = {git = "https://github.com/Mattwmaster58/playwright_stealth"}
aiohttp = "^3.11.7"
aiofiles = "^24.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.4"
pre-commit = "^4.0.1"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
preview = true
fixable = ["ALL"]
select = ['E', 'F', 'W', 'I']

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
